% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_from_file}
\alias{initialize_from_file}
\title{Initialize tables used to store a simulation's state from files}
\usage{
initialize_from_file(
  path,
  current_date,
  history_start_date,
  future_end_date,
  path_to_demand,
  path_to_inventory,
  path_to_products,
  seed = 512
)
}
\arguments{
\item{path}{Path to the directory in which the gym's state is stored}

\item{current_date}{The date from which the simulation will pick up}

\item{history_start_date}{The first date in the historical data, used to
create a sales history so that there is training data for forecast models}

\item{future_end_date}{The final date in the future for which data will be
created; the simulation can't update further into the future}

\item{path_to_demand}{File path to the parquet file used as basis to derive
demand and sales from}

\item{path_to_inventory}{File path to the parquet file used as basis to
derive inventory from}

\item{path_to_products}{File path to the parquet file used as basis to derive
product information including prices from}
}
\description{
In contrast to \code{initialize()}, this function creates the tables by taking
the required information from three parquet files that provide the necessary
information for demand, inventory, and products.
}
\details{
A side effect of this function is the creation of most of the tables listed
in \code{list_tables()} in the directory specified with \code{path}.
}
