% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{list_tables}
\alias{list_tables}
\title{List tables managed by \code{shopr}}
\usage{
list_tables()
}
\value{
A list of lists, each of which represent one table managed by \code{shopr}
to maintain the simulation's current state. Each table definition comes
with \code{arrow} partitioning and schema.
}
\description{
List tables managed by \code{shopr}
}
