% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{open_table}
\alias{open_table}
\title{Open a table managed by \code{shopr}}
\usage{
open_table(table, path)
}
\arguments{
\item{table}{Name of the table that should be opened, as defined in
\code{shopr}'s internal schema (see \code{shopr::list_tables()}).}

\item{path}{Path to the directory in which the gym's state is stored}
}
\description{
Wrapper around \code{arrow::open_dataset()} to open a table managed by \code{shopr} as
data set.
}
