% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record_profit_and_loss}
\alias{record_profit_and_loss}
\title{Write tables that record profit and loss for the current date}
\usage{
record_profit_and_loss(current_date, path, verbose = FALSE)
}
\description{
Write tables that record profit and loss for the current date
}
