% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_profit_and_loss}
\alias{report_profit_and_loss}
\title{Report profit and loss for a range of dates}
\usage{
report_profit_and_loss(from_date, to_date, path)
}
\description{
Report profit and loss for a range of dates
}
