% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_clean_datetime.R
\name{shopr_clean_datetime}
\alias{shopr_clean_datetime}
\title{Clean datetime}
\usage{
shopr_clean_datetime(datetime)
}
\arguments{
\item{datetime}{vector of datetimes with class character or POSIXct}
}
\value{
vector of datetimes in character format like yyyy-mm-ddTHH:MM:SS-00:00
}
\description{
Helper function to validate datetimes and convert them to yyyy-mm-ddTHH:MM:SS-00:00
}
