% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_clean_fields.R
\name{shopr_clean_fields}
\alias{shopr_clean_fields}
\title{Clean fields}
\usage{
shopr_clean_fields(fields, resource)
}
\arguments{
\item{fields}{vector of fields}

\item{resource}{'orders' or 'products'}
}
\value{
string of clean field names separated by a comma
}
\description{
Helper function to validate validate and clean fields
}
