% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_clean_financial_status.R
\name{shopr_clean_financial_status}
\alias{shopr_clean_financial_status}
\title{Clean financial status}
\usage{
shopr_clean_financial_status(financial_status)
}
\arguments{
\item{financial_status}{one of {'any', 'authorized', 'pending', 'paid', 'refunded', 'voided'}}
}
\value{
string of clean financial_status
}
\description{
Helper function to validate and clean financial status parameter
}
