% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_clean_fulfillment_status.R
\name{shopr_clean_fulfillment_status}
\alias{shopr_clean_fulfillment_status}
\title{Clean fulfillment status}
\usage{
shopr_clean_fulfillment_status(fulfillment_status)
}
\arguments{
\item{fulfillment_status}{one of {'any', 'shipped', 'partial', 'unshipped'}}
}
\value{
string of clean financial_status
}
\description{
Helper function to validate and clean fulfillment status parameter
}
