% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_clean_ids.R
\name{shopr_clean_ids}
\alias{shopr_clean_ids}
\title{Clean ids}
\usage{
shopr_clean_ids(ids, resource, since_id = NULL)
}
\arguments{
\item{ids}{vector of ids or string of comma separated ids}

\item{resource}{one of {'orders', 'products', 'inventory'}}

\item{since_id}{optional since id value}
}
\value{
string of clean ids parameter
}
\description{
Helper function to validate and clean ids parameter
}
