% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_clean_order_status.R
\name{shopr_clean_order_status}
\alias{shopr_clean_order_status}
\title{Clean order status}
\usage{
shopr_clean_order_status(order_status)
}
\arguments{
\item{order_status}{one of {'open', 'closed', 'any'}}
}
\value{
string of clean order_status
}
\description{
Helper function to validate and clean order status parameter
}
