% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_clean_published_status.R
\name{shopr_clean_published_status}
\alias{shopr_clean_published_status}
\title{Clean published status}
\usage{
shopr_clean_published_status(published_status)
}
\arguments{
\item{published_status}{one of {'any', 'published', 'unpublished'}}
}
\value{
string of clean published_status
}
\description{
Helper function to validate and clean published status parameter
}
