% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_get_inventory_items.R
\name{shopr_get_inventory_items}
\alias{shopr_get_inventory_items}
\title{Get Inventory Items}
\usage{
shopr_get_inventory_items(shopURL, APIKey, APIPassword,
  APIVersion = NULL, max_pages = Inf, limit_per_page = 250L, ids,
  since_id = 0L, verbose = FALSE)
}
\arguments{
\item{shopURL}{shop URL (e.g. 'https://superstore-1.myshopify.com')}

\item{APIKey}{API key}

\item{APIPassword}{API password}

\item{APIVersion}{API version (default = \code{NULL} -> use the latest version)}

\item{max_pages}{maximum pages of records to return (i.e. the maximum number of HTTP GET requests to make). (default
= \code{Inf})}

\item{limit_per_page}{maximum number of records to return per page (i.e. per HTTP GET request). Should be in the
range [1, 250]. (default = 250)}

\item{ids}{vector of inventory item ids like c(123, 456) (REQUIRED)}

\item{since_id}{only return inventory items with id > since_id (default = 0)}

\item{verbose}{should progress messages be printed? (default = \code{FALSE})}
}
\value{
data.table of inventory items
}
\description{
Retrieve inventory items for a shop
}
\details{
An inventory item represents the physical good available to be shipped to a customer. It holds essential information
about the physical good, including its SKU and whether its inventory is tracked.

There is a 1:1 relationship between a product variant and an inventory item. Each product variant includes the ID of
its related inventory item. You can use the inventory item ID to query the InventoryLevel resource to retrieve
inventory information.

Use the InventoryItem resource together with the InventoryLevel and Location resources to manage a store's inventory
across multiple locations.
}
\references{
\url{https://help.shopify.com/en/api/reference/inventory/inventoryitem}
}
