% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_get_inventory_levels.R
\name{shopr_get_inventory_levels}
\alias{shopr_get_inventory_levels}
\title{Get Inventory Levels}
\usage{
shopr_get_inventory_levels(
  shopURL,
  APIKey,
  APIPassword,
  APIVersion = NULL,
  max_pages = Inf,
  limit_per_page = 250L,
  inventory_item_ids = NULL,
  location_ids = NULL,
  updated_at_min = NULL,
  verbose = FALSE
)
}
\arguments{
\item{shopURL}{shop URL (e.g. 'https://superstore-1.myshopify.com')}

\item{APIKey}{API key}

\item{APIPassword}{API password}

\item{APIVersion}{API version (default = \code{NULL} -> use the latest version)}

\item{max_pages}{max number of pages of orders to return. (default = \code{Inf})}

\item{limit_per_page}{max number of results to return per request. Should be in the range [1, 250]. (default = 250)}

\item{inventory_item_ids}{inventory item ids. At least one of \code{inventory_item_ids} and \code{location_ids} must
be provided. (default = \code{NULL})}

\item{location_ids}{location ids. At least one of \code{inventory_item_ids} and \code{location_ids} must be provided.
(default = \code{NULL})}

\item{updated_at_min}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{verbose}{should progress messages be printed? (default = \code{FALSE})}
}
\value{
data.table of inventory levels
}
\description{
Retrieve inventory levels for a shop
}
\details{
An inventory level represents the available quantity of an inventory item at a specific location.

Each inventory level belongs to one inventory item and has one location. For every location where an inventory item
is available, there's an inventory level that represents the inventory item's quantity at that location:
}
\references{
\url{https://help.shopify.com/en/api/reference/inventory/inventorylevel#index}
}
