% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_get_latest_api_version.R
\name{shopr_get_latest_api_version}
\alias{shopr_get_latest_api_version}
\title{Get Latest Stable API Version}
\usage{
shopr_get_latest_api_version()
}
\value{
character indicating the latest stable API version
}
\description{
Retrieve the latest stable api version
}
\details{
Shopify releases a new API version every 3 months at the beginning of the quarter. Version names are date based to
be meaningful and semantically unambiguous (for example, 2020-01). Stable versions are released at 5pm UTC.
}
\references{
\url{https://help.shopify.com/en/api/versioning#the-api-version-release-schedule}
}
