% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_get_locations.R
\name{shopr_get_locations}
\alias{shopr_get_locations}
\title{Get Locations}
\usage{
shopr_get_locations(
  shopURL,
  APIKey,
  APIPassword,
  APIVersion = NULL,
  verbose = FALSE
)
}
\arguments{
\item{shopURL}{shop url (e.g. 'https://superstore-1.myshopify.com')}

\item{APIKey}{API key}

\item{APIPassword}{API password}

\item{APIVersion}{API version}

\item{verbose}{should progress messages be printed? (default FALSE)}
}
\value{
data.table of locations
}
\description{
Retrieve the locations associated with a shop
}
\details{
A location represents a geographical location where stores, pop-up stores, headquarters, and warehouses exist.
The Location resource can be used to track sales, manage inventory, and configure the tax rates to apply at checkout.
}
\examples{
shopr_get_locations(
  shopURL = "https://superstore-1.myshopify.com",
  APIKey = "abc123",
  APIPassword = "123abc"
)
}
\references{
\url{https://help.shopify.com/en/api/reference/inventory/location#index}
}
