% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_get_orders_count.R
\name{shopr_get_orders_count}
\alias{shopr_get_orders_count}
\title{Get Orders Count}
\usage{
shopr_get_orders_count(shopURL, APIKey, APIPassword, APIVersion = NULL,
  ids = NULL, since_id = NULL, created_at_min = NULL,
  created_at_max = NULL, updated_at_min = NULL,
  updated_at_max = NULL, processed_at_min = NULL,
  processed_at_max = NULL, status = "any", financial_status = NULL,
  fulfillment_status = NULL, verbose = FALSE)
}
\arguments{
\item{shopURL}{shop URL (e.g. 'https://superstore-1.myshopify.com')}

\item{APIKey}{API key}

\item{APIPassword}{API password}

\item{APIVersion}{API version (default = \code{NULL} -> use the latest version)}

\item{ids}{vector of order ids like c(123, 456) (default = \code{NULL})}

\item{since_id}{only consider orders with id > since_id (default = 0)}

\item{created_at_min}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{created_at_max}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{updated_at_min}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{updated_at_max}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{processed_at_min}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{processed_at_max}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{status}{order status. Should be \code{NULL} or one of \{'open', 'closed', 'any'\}. (default = \code{'any'})}

\item{financial_status}{financial status. Should be \code{NULL} or one of \{'any', 'authorized', 'pending', 'paid',
'refunded', 'voided'\}. (default = \code{NULL})}

\item{fulfillment_status}{fulfillment status. Should be \code{NULL} or one of \{'any', 'shipped', 'partial',
'unshipped'\}. (default = \code{NULL})}

\item{verbose}{should progress messages be printed? (default = \code{FALSE})}
}
\value{
number of orders
}
\description{
Retrieve the number of orders for a shop satisfying some criteria
}
\references{
\url{https://help.shopify.com/en/api/reference/orders/order#count}
}
