% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_get_products_count.R
\name{shopr_get_products_count}
\alias{shopr_get_products_count}
\title{Get Products Count}
\usage{
shopr_get_products_count(shopURL, APIKey, APIPassword, APIVersion = NULL,
  ids = NULL, since_id = NULL, title = NULL, vendor = NULL,
  handle = NULL, product_type = NULL, collection_id = NULL,
  created_at_min = NULL, created_at_max = NULL,
  updated_at_min = NULL, updated_at_max = NULL,
  published_at_min = NULL, published_at_max = NULL,
  published_status = NULL)
}
\arguments{
\item{shopURL}{shop URL (e.g. 'https://superstore-1.myshopify.com')}

\item{APIKey}{API key}

\item{APIPassword}{API password}

\item{APIVersion}{API version (default = \code{NULL} -> use the latest version)}

\item{ids}{vector of product ids like c(123, 456) (default = \code{NULL})}

\item{since_id}{only consider products with id > since_id (default = 0)}

\item{title}{product title (default = \code{NULL})}

\item{vendor}{product vendor (default = \code{NULL})}

\item{handle}{product handle (default = \code{NULL})}

\item{product_type}{product type (default = \code{NULL})}

\item{collection_id}{collection id (default = \code{NULL})}

\item{created_at_min}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{created_at_max}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{updated_at_min}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{updated_at_max}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{published_at_min}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{published_at_max}{POSIXct datetime or a string with format like '2014-04-25T16:15:47-04:00' (default =
\code{NULL})}

\item{published_status}{published status. Should be NULL or one of \{'published', 'unpublished', 'any'\}. (default =
\code{NULL})}
}
\value{
number of products
}
\description{
Retrieve the number of products for a shop satisfying some criteria
}
\references{
\url{https://help.shopify.com/en/api/reference/products/product#count}
}
