% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_make_requests.R
\name{shopr_make_requests}
\alias{shopr_make_requests}
\title{Make Requests}
\usage{
shopr_make_requests(
  requestURL,
  params,
  pagesN,
  maxPages = Inf,
  APIKey,
  APIPassword,
  verbose = FALSE
)
}
\arguments{
\item{requestURL}{e.g. https://abc.myshopify.com/admin/api/2020-01/orders.json?limit=250&since_id=0}

\item{params}{list of request parameters that must include 'limit'}

\item{pagesN}{expected number of pages to be returned (used for printing progress and initializing response list)}

\item{maxPages}{maximum number of pages to return (default = \code{Inf})}

\item{APIKey}{API key}

\item{APIPassword}{API password}

\item{verbose}{should progress messages be printed? (default = \code{FALSE})}
}
\value{
list of http responses
}
\description{
Generic helper method for making API requests that return paginated results using cursor-based pagination
}
\references{
\url{https://help.shopify.com/en/api/guides/paginated-rest-results}
}
