% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_parse_call_limit.R
\name{shopr_parse_call_limit}
\alias{shopr_parse_call_limit}
\title{Prase Call Limit}
\usage{
shopr_parse_call_limit(xShopifyAPICallLimit)
}
\arguments{
\item{xShopifyAPICallLimit}{header field xShopifyAPICallLimit from recent response}
}
\value{
list of {bucketCalls, bucketSize}
}
\description{
Helper function to parse the api call limit given the latest response
}
