% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopr_utcoffset_to_posixct.R
\name{shopr_UTCOffset_to_POSIXct}
\alias{shopr_UTCOffset_to_POSIXct}
\title{UTCOffset To POSIXct}
\usage{
shopr_UTCOffset_to_POSIXct(x, tz = "UTC")
}
\arguments{
\item{x}{character vector of UTC offset strings like '2014-04-25T16:15:47-04:00'}

\item{tz}{timezone (default = 'UTC')}
}
\value{
vector of POSIXct
}
\description{
Convert a UTC offset string (e.g. '2014-04-25T16:15:47-04:00') to POSIXct
}
\examples{
shopr_UTCOffset_to_POSIXct(x = c('2019-01-01T13:15:00-04:00', '2019-01-01T13:15:00+04:00'))
shopr_UTCOffset_to_POSIXct(x = c('2019-01-01T13:15:00-04:00', '2019-01-01T13:15:00+04:00'), tz = 'America/Chicago')
}
\references{
\url{https://en.wikipedia.org/wiki/UTC_offset}
}
