% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update the gym's state by one day}
\usage{
update(path)
}
\arguments{
\item{path}{Path to the directory in which the gym's state is stored}
}
\value{
Returns the date of the current iteration invisibly
}
\description{
This function updates several tables stored at \code{path} as side effect,
including the \code{settings}, \code{inventory} and \code{sales} tables.
}
