% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{write_table}
\alias{write_table}
\title{Write a table managed by \code{shopr}}
\usage{
write_table(dataset, table, path)
}
\arguments{
\item{dataset}{A data frames or \code{arrow} data set}

\item{table}{Name of the table that should be written to, as defined in
\code{shopr}'s internal schema (see \code{shopr::list_tables()}).}

\item{path}{Path to the directory in which the gym's state is stored}
}
\value{
TRUE, invisibly
}
\description{
Write a table managed by \code{shopr}
}
