% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{write_tables}
\alias{write_tables}
\title{Write tables managed by \code{shopr}}
\usage{
write_tables(named_datasets, path)
}
\arguments{
\item{named_datasets}{A named list of data frames or \code{arrow} data sets}

\item{path}{Path to the directory in which the gym's state is stored}
}
\value{
TRUE, invisibly
}
\description{
Wrapper around \code{write_table} that writes a named list of tables by looking up
the table's names in \code{shopr}'s table schema.
}
